@@ -1,8 +1,9 @@
 import React, {useState} from 'react';
+import { toast } from 'react-toastify';
+import theme from '../theme.jsx';
 import {BrowserView, MobileView} from "react-device-detect";
-import { useTheme } from '@material-ui/core/styles';
 
-import {Link} from 'react-router-dom';
+import { useNavigate, Link } from "react-router-dom";
 import ReactGA from 'react-ga4';
 
 import {
@@ -21,7 +22,7 @@ import {
 	IconButton,
 	Divider,
 	LinearProgress, 
-} from '@material-ui/core'
+} from '@mui/material'
 
 import { 
 	MeetingRoom as MeetingRoomIcon, 
@@ -29,28 +30,28 @@ import {
 	Settings as SettingsIcon, 
 	Notifications as NotificationsIcon, 
 	Home as HomeIcon, 
-	Polymer as PolymerIcon, 
 	Apps as AppsIcon, 
 	Description as DescriptionIcon,
 	EmojiObjects as EmojiObjectsIcon,
   Business as BusinessIcon,
-} from '@material-ui/icons';
+
+	Polyline as PolylineIcon, 
+} from '@mui/icons-material';
 
 import {
 	Analytics as AnalyticsIcon,
 	Lightbulb as LightbulbIcon,
 } from "@mui/icons-material";
 
-import { useAlert } from "react-alert";
-
 import SearchField from '../components/Searchfield.jsx'
+
 const hoverColor = "#f85a3e"
 const hoverOutColor = "#e8eaf6"
 
 const Header = props => {
 const { globalUrl, setNotifications, notifications, isLoggedIn, removeCookie, homePage, userdata, serverside, } = props;
-	const theme = useTheme();
-	const alert = useAlert()
+	//const theme = useTheme();
+	//const alert = useAlert()
 
 
 	const [HomeHoverColor, setHomeHoverColor] = useState(hoverOutColor);
@@ -61,6 +62,7 @@ const { globalUrl, setNotifications, notifications, isLoggedIn, removeCookie, ho
   const [anchorEl, setAnchorEl] = React.useState(null);
   const [anchorElAvatar, setAnchorElAvatar] = React.useState(null);
   const [subAnchorEl, setSubAnchorEl] = React.useState(null);
+  let navigate = useNavigate();
 
 
   const handleClick = (event) => {
@@ -99,7 +101,7 @@ const { globalUrl, setNotifications, notifications, isLoggedIn, removeCookie, ho
 				setNotifications([])
 				handleClose()
 			} else {
-				alert.error("Failed dismissing notifications. Please try again later.")
+				toast("Failed dismissing notifications. Please try again later.")
 			}
 		})
 		.catch(error => {
@@ -129,7 +131,7 @@ const { globalUrl, setNotifications, notifications, isLoggedIn, removeCookie, ho
 				console.log("NEW NOTIFICATIONS: ", newNotifications)
 				setNotifications(newNotifications)
 			} else {
-				alert.error("Failed dismissing notification. Please try again later.")
+				toast("Failed dismissing notification. Please try again later.")
 			}
 		})
 		.catch(error => {
@@ -209,48 +211,49 @@ const { globalUrl, setNotifications, notifications, isLoggedIn, removeCookie, ho
 
 	const notificationWidth = 300
 	const imagesize = 22;
-  const boxColor = "#86c142";
+  	const boxColor = "#86c142";
+
 	const NotificationItem = (props) => {
 		const {data} = props
 
-    var image = "";
-    var orgName = "";
-    var orgId = "";
-    if (userdata.orgs !== undefined) {
-      const foundOrg = userdata.orgs.find((org) => org.id === data["org_id"]);
-      if (foundOrg !== undefined && foundOrg !== null) {
-        //position: "absolute", bottom: 5, right: -5,
-        const imageStyle = {
-          width: imagesize,
-          height: imagesize,
-          pointerEvents: "none",
-          marginLeft: data.creator_org !== undefined && data.creator_org.length > 0 ? 20 : 0,
-          borderRadius: 10,
-          border: foundOrg.id === userdata.active_org.id ? `3px solid ${boxColor}` : null,
-          cursor: "pointer",
-          marginRight: 10,
-        };
+		var image = "";
+		var orgName = "";
+		var orgId = "";
+		if (userdata.orgs !== undefined) {
+		  const foundOrg = userdata.orgs.find((org) => org.id === data["org_id"]);
+		  if (foundOrg !== undefined && foundOrg !== null) {
+			//position: "absolute", bottom: 5, right: -5,
+			const imageStyle = {
+			  width: imagesize,
+			  height: imagesize,
+			  pointerEvents: "none",
+			  marginLeft: data.creator_org !== undefined && data.creator_org.length > 0 ? 20 : 0,
+			  borderRadius: 10,
+			  border: foundOrg.id === userdata.active_org.id ? `3px solid ${boxColor}` : null,
+			  cursor: "pointer",
+			  marginRight: 10,
+			};
 
-        image =
-          foundOrg.image === "" ? (
-            <img
-              alt={foundOrg.name}
-              src={theme.palette.defaultImage}
-              style={imageStyle}
-            />
-          ) : (
-            <img
-              alt={foundOrg.name}
-              src={foundOrg.image}
-              style={imageStyle}
-              onClick={() => {}}
-            />
-          );
+			image =
+			  foundOrg.image === "" ? (
+				<img
+				  alt={foundOrg.name}
+				  src={theme.palette.defaultImage}
+				  style={imageStyle}
+				/>
+			  ) : (
+				<img
+				  alt={foundOrg.name}
+				  src={foundOrg.image}
+				  style={imageStyle}
+				  onClick={() => {}}
+				/>
+			  );
 
-        orgName = foundOrg.name;
-        orgId = foundOrg.id;
-      }
-    }
+			orgName = foundOrg.name;
+			orgId = foundOrg.id;
+		  }
+		}
 
 		return (
 			<Paper style={{backgroundColor: theme.palette.surfaceColor, width: notificationWidth, padding: 25, borderBottom: "1px solid rgba(255,255,255,0.4)"}}>
@@ -260,7 +263,7 @@ const { globalUrl, setNotifications, notifications, isLoggedIn, removeCookie, ho
 				{data.reference_url !== undefined && data.reference_url !== null && data.reference_url.length > 0 ?
 					<Link to={data.reference_url} style={{color: "#f86a3e", textDecoration: "none",}}>
 						<Typography variant="body1">
-							{data.title}
+							{data.title} ({data.amount})
 						</Typography >
 					</Link>
 				: 
@@ -274,7 +277,7 @@ const { globalUrl, setNotifications, notifications, isLoggedIn, removeCookie, ho
 					: 
 					null
 				}
-				<Typography variant="body2">
+				<Typography variant="body2" style={{maxHeight: 200, overflowX: "hidden", overflowY: "auto", }}>
 					{data.description}
 				</Typography >
 				{/*data.tags !== undefined && data.tags !== null && data.tags.length > 0 ? 
@@ -404,9 +407,9 @@ const { globalUrl, setNotifications, notifications, isLoggedIn, removeCookie, ho
 				setTimeout(() => {
 					window.location.reload()
 				}, 2000)
-				alert.success("Successfully changed active organization - refreshing!")
+				toast("Successfully changed active organization - refreshing!")
 			} else {
-				alert.error("Failed changing org: ", responseJson.reason)
+				toast("Failed changing org: ", responseJson.reason)
 			}
 		})
 		.catch(error => {
@@ -518,6 +521,7 @@ const { globalUrl, setNotifications, notifications, isLoggedIn, removeCookie, ho
 	}
 
 	// Handle top bar or something
+  const defaultTop = isCloud ? 0 : 7
   const loginTextBrowser = !isLoggedIn ? 
     <div style={{display: "flex", minWidth: 1250, maxWidth: 1250, margin: "auto", textAlign: "center",}}>
 			<div style={{display: "flex", flex: 1, }}>
@@ -610,7 +614,7 @@ const { globalUrl, setNotifications, notifications, isLoggedIn, removeCookie, ho
 
 				</List>
 			</div>
-			<div style={{flex: 1, marginTop: 0,}}>
+			<div style={{flex: 1, marginTop: 10,}}>
 				<SearchField serverside={serverside} userdata={userdata} />
 			</div>
 			<div style={{flex: 1, display: "flex", flexDirection: "row-reverse"}}>
@@ -668,7 +672,7 @@ const { globalUrl, setNotifications, notifications, isLoggedIn, removeCookie, ho
 			</div>
 	    </div>
     	: 
-		<div style={{display: "flex", backgroundColor: "#1f2023",}}>
+		<div style={{display: "flex", }}>
 			<div style={{minWidth: 1250, maxWidth: 1250, display: "flex", margin: "auto", }}>
 				<div style={{flex: 1, flexDirection: "row"}}>
 					<List style={{height: 56, marginTop: "auto", marginBottom: "auto", display: "flex", flexDirect: "row", alignItems: "baseline", maxWidth: 340, }} component="nav">
@@ -690,9 +694,9 @@ const { globalUrl, setNotifications, notifications, isLoggedIn, removeCookie, ho
 								<Link to="/workflows" style={hrefStyle}>
 									<div onMouseOver={handleSoarHover} onMouseOut={handleSoarHoverOut} style={{color: SoarHoverColor, cursor: "pointer", display: "flex"}}>
 										{/*
-										<PolymerIcon style={{marginRight: "5px"}} />
+										<PolylineIcon style={{marginRight: "5px"}} />
 										*/}
-										<span style={{marginTop: 0, marginRight: 8, }}>Workflows</span>
+										<Typography style={{marginTop: defaultTop, marginRight: 8, }}>Workflows</Typography>
 									</div> 
 								</Link>
       	 			</ListItem>
@@ -702,7 +706,7 @@ const { globalUrl, setNotifications, notifications, isLoggedIn, removeCookie, ho
 											{/*
 											<AppsIcon style={{marginRight: "5px"}} />
 											*/}
-											<span style={{marginTop: 0, marginRight: 5,  }}>Apps</span>
+											<Typography style={{marginTop: defaultTop, marginRight: 5,  }}>Apps</Typography>
 										</div>
 								</Link>
       	 			</ListItem>
@@ -719,7 +723,7 @@ const { globalUrl, setNotifications, notifications, isLoggedIn, removeCookie, ho
 											{/*
 											<DescriptionIcon style={{marginRight: "5px"}} />
 											*/}
-											<span style={{marginTop: 0,}}>Docs</span>
+											<Typography style={{marginTop: defaultTop,}}>Docs</Typography>
 										</div>
 								</Link>
       	 			</ListItem>
@@ -742,7 +746,7 @@ const { globalUrl, setNotifications, notifications, isLoggedIn, removeCookie, ho
 						*/}
 					</List>
 				</div>
-				<div style={{flex: 1, marginTop: 0,}}>
+				<div style={{flex: 1, marginTop: 10,}}>
 					<SearchField serverside={serverside} userdata={userdata} />
 				</div>
 				<div style={{flex: 1, display: "flex", flexDirection: "row-reverse"}}>
@@ -793,8 +797,6 @@ const { globalUrl, setNotifications, notifications, isLoggedIn, removeCookie, ho
 								</ListItem>
 						: null*/}
 
-						
-
 						{userdata === undefined || userdata.orgs === undefined || userdata.orgs === null || userdata.orgs.length <= 1 ? 
 							null
 						:
@@ -811,7 +813,7 @@ const { globalUrl, setNotifications, notifications, isLoggedIn, removeCookie, ho
 									}}
 									MenuProps={{
 										style: {
-											zIndex: 10002,
+											zIndex: 15000,
 										},
 									}}
 									style={{
@@ -885,7 +887,7 @@ const { globalUrl, setNotifications, notifications, isLoggedIn, removeCookie, ho
 
 												<Tooltip color="primary" title={parsedTitle} placement="left">
 													<div style={{display: "flex"}}>
-														<Typography variant="body2" style={{borderRadius: theme.palette.borderRadius, float: "left", margin: "0 0 0 0", marginRight: 25, }}>{regiontag}</Typography> {image} <span style={{marginLeft: 8}}>{data.name}</span> 
+														{isCloud?<Typography variant="body2" style={{borderRadius: theme.palette.borderRadius, float: "left", margin: "0 0 0 0", marginRight: 25, }}>{regiontag}</Typography>:null} {image} <span style={{marginLeft: 8}}>{data.name}</span> 
 													</div>
 												</Tooltip>
 											</MenuItem>
@@ -895,13 +897,32 @@ const { globalUrl, setNotifications, notifications, isLoggedIn, removeCookie, ho
 							</span>
 							}
 
+						{/* Show on cloud, if not suborg and if not customer/pov/internal */}
+						{isCloud && (userdata.org_status === undefined || userdata.org_status === null || userdata.org_status.length === 0) ? 
+							<ListItem style={{textAlign: "center", marginLeft: 0, marginRight: 7, marginTop: 3, }}>
+								<Link to ="/pricing?tab=cloud&highlight=true" style={hrefStyle}>
+									<Button variant="contained" color="primary" style={{textTransform: "none"}} onClick={() => {
+										ReactGA.event({
+											category: "header",
+											action: "pricing_upgrade_click",
+											label: "",
+										})
+									}}>
+										Upgrade	
+									</Button>
+								</Link>
+							</ListItem>
+						: null}
+
 						{userdata === undefined || userdata.app_execution_limit === undefined || userdata.app_execution_usage === undefined || userdata.app_execution_usage < 1000 ? 
 							null
 							:
 							<Tooltip title={`Amount of executions left: ${userdata.app_execution_usage} / ${userdata.app_execution_limit}. When the limit is reached, you can still use Gsoc2 normally, but your Workflow triggers may stop working. Reach out to support@gsoc2r.io to extend this limit.`}>
-								<div style={{maxHeight: 30, minHeight: 30, padding: 8, textAlign: "center", cursor: "pointer", borderRadius: theme.palette.borderRadius, marginRight: 10, marginTop: 12, backgroundColor: theme.palette.surfaceColor, minWidth: 60, maxWidth: 60,  }} onClick={() => {
+								<div style={{maxHeight: 30, minHeight: 30, padding: 8, textAlign: "center", cursor: "pointer", borderRadius: theme.palette.borderRadius, marginRight: 10, marginTop: 12, backgroundColor: theme.palette.surfaceColor, minWidth: 60, maxWidth: 60, border: userdata.app_execution_usage/userdata.app_execution_limit >= 0.9 ? "#f86a3e" : null, }} onClick={() => {
+										console.log(userdata.appe_execution_usage/userdata.app_execution_limit)
 										if (window.drift !== undefined) {
 											window.drift.api.startInteraction({ interactionId: 326905 })
+											navigate("/pricing")
 										} else {
 											console.log("Couldn't find drift in window.drift and not .drift-open-chat with querySelector: ", window.drift)
 										}
@@ -1011,16 +1032,16 @@ const { globalUrl, setNotifications, notifications, isLoggedIn, removeCookie, ho
 	const loadedCheck = 
 		<div style={{minHeight: 68}}>
 			<BrowserView>
-      	{loginTextBrowser}
+      			{loginTextBrowser}
 			</BrowserView>
 			<MobileView>
-      	{loginTextMobile}
+      			{loginTextMobile}
 			</MobileView>
 		</div>
     // <div style={{backgroundImage: "linear-gradient(-90deg,#342f78 0,#29255e 50%,#1b1947 100%"}}>
   	return (
-    	<div style={{backgroundColor: props.color === "undefined" ? "inherit" : props.color}}>
-				{loadedCheck}
+    	<div style={{backgroundColor: theme.palette.backgroundColor, }}>
+			{loadedCheck}
 	    </div>
   )
 }
