@@ -1,11 +1,12 @@
 import React, { useState, useEffect } from 'react';
 import ReactGA from 'react-ga4';
 import WelcomeForm2 from "../components/WelcomeForm2.jsx";
-import Stepper from "@material-ui/core/Stepper";
-import Step from "@material-ui/core/Step";
-import StepLabel from "@material-ui/core/StepLabel";
 import AppFramework from "../components/AppFramework.jsx";
-import ArrowBackIosIcon from '@mui/icons-material/ArrowBackIos';
+import {isMobile} from "react-device-detect";
+import {
+	ArrorForwardIos as ArrowForwardIosIcon,
+} from '@mui/icons-material';
+
 import { 
 	Grid, 
 	Container,
@@ -16,37 +17,37 @@ import {
 	Card,
 	CardContent,
 	CardActionArea,
+	Stepper,
+	Step,
+	StepLabel,
 } from '@mui/material';
 import theme from '../theme.jsx';
 import { useNavigate, Link } from "react-router-dom";
+import ArrowBackIosIcon from '@mui/icons-material/ArrowBackIos';
 import Drift from "react-driftjs";
 
 const Welcome = (props) => {
-    const { globalUrl, surfaceColor, newColor, mini, inputColor, userdata, isLoggedIn, isLoaded, serverside } = props;
+    const { globalUrl, surfaceColor, newColor, mini, inputColor, userdata, isLoggedIn, isLoaded, serverside, checkLogin } = props;
     const [skipped, setSkipped] = React.useState(new Set());
     const [inputUsecase, setInputUsecase] = useState({});
   	const [frameworkData, setFrameworkData] = useState(undefined);
   	const [discoveryWrapper, setDiscoveryWrapper] = useState(undefined);
     const [activeStep, setActiveStep] = React.useState(1);
   	const [apps, setApps] = React.useState([]);
-		const [defaultSearch, setDefaultSearch] = React.useState("")
-		const [selectionOpen, setSelectionOpen] = React.useState(false)
-		const [showWelcome, setShowWelcome] = React.useState(false)
+	const [defaultSearch, setDefaultSearch] = React.useState("")
+	const [selectionOpen, setSelectionOpen] = React.useState(false)
+	const [showWelcome, setShowWelcome] = React.useState(false)
   	const [usecases, setUsecases] = React.useState([]);
   	const [workflows, setWorkflows] = React.useState([]);
 
-		let navigate = useNavigate();
-    //if (serverside === false && isLoaded === true && isLoggedIn === false) {
-		//	console.log("Redirecting to login?")
-		//	console.log(window.location.pathname)
-		//	console.log(window.location)
-		//	navigate(`/login?view=${window.location.pathname}${window.location.search}`)
-		//}
+	let navigate = useNavigate();
+	useEffect(() => {
+		if (checkLogin !== undefined) {
+			checkLogin()
+		}
+	}, [activeStep])
 
-  	const isCloud =
-			window.location.host === "localhost:3002" ||
-			window.location.host === "gsoc2r.io";
-
+  	const isCloud = window.location.host === "localhost:3002" || window.location.host === "gsoc2r.io";
 		const [steps, setSteps] = useState([
 			"Help us get to know you", 
 			"Find your Apps", 
@@ -191,9 +192,9 @@ const Welcome = (props) => {
 					setFrameworkData({})
 
 					if (responseJson.reason !== undefined) {
-						//alert.error("Failed loading: " + responseJson.reason)
+						//toast("Failed loading: " + responseJson.reason)
 					} else {
-						//alert.error("Failed to load framework for your org.")
+						//toast("Failed to load framework for your org.")
 					}
 				} else {
 					setFrameworkData(responseJson)
@@ -268,23 +269,22 @@ const Welcome = (props) => {
 			getApps()
 			getAvailableWorkflows() 
 
-			if (
-				window.location.search !== undefined &&
-				window.location.search !== null
-			) {
+			if (window.location.search !== undefined && window.location.search !== null) {
 				const urlSearchParams = new URLSearchParams(window.location.search);
 				const params = Object.fromEntries(urlSearchParams.entries());
 				const foundTab = params["tab"];
-				if (foundTab !== null && foundTab !== undefined && !isNaN(foundTab)) {
+
+				// Make foundtab into a number
+				console.log("foundTab: ", foundTab, activeStep+1)
+				if (foundTab !== null && foundTab !== undefined && !isNaN(foundTab) && foundTab >= 1 && foundTab <= 3) {
+					console.log("SETTING TAB TO: ", foundTab)
+
 					setShowWelcome(true)
 					if (foundTab === 3 || foundTab === "3") {
 						handleSetSearch(usecaseButtons[0].name, usecaseButtons[0].usecase)
 					}
 
 					setActiveStep(foundTab-1)
-				} else { 
-    			//navigate(`/welcome?tab=1`)
-    			navigate(`/welcome?tab=2`)
 				}
 			}
 		}, [])
@@ -297,44 +297,47 @@ const Welcome = (props) => {
 			flex: 1, 
 			padding: 0, 
 			textAlign: "center",
-			maxWidth: 300,
-			minWidth: 300,
+			maxWidth: isMobile ? null : 275,
+			minWidth: isMobile ? null : 275, 
 			backgroundColor: theme.palette.surfaceColor,
 			color: "white",
 			borderRadius: theme.palette.borderRadius,
 		}
 
 		const actionObject = {
-			padding: "35px", 
-			maxHeight: 300,
-			minHeight: 300,
+			padding: "25px", 
+			maxHeight: 280,
+			minHeight: 280,
 			borderRadius: theme.palette.borderRadius,
 		}
 
 		const imageStyle = {
-			width: 150, 
+			width: 70, 
 			// height: 150, 
 			// margin: "auto", 
 			// marginTop: 10, 
-			borderRadius: 75, 
+			marginBottom: 18,
+			// borderRadius: 75, 
 			objectFit: "scale-down",
 		}
 		const buttonStyle = { 
-			borderRadius: 8, 
+			borderRadius: 200, 
 			height: 51, 
-			width: 464, 
+			width: isMobile ? 300 : 464, 
 			fontSize: 16, 
-			background: "linear-gradient(89.83deg, #FF8444 0.13%, #F2643B 99.84%)", 
+			// background: "linear-gradient(89.83deg, #FF8444 0.13%, #F2643B 99.84%)", 
+			background: "linear-gradient(90deg, #F86744 0%, #F34475 100%)",
 			padding: "16px 24px", 
-			top: 75, 
+			top: 105, 
 			margin: "auto",
 			itemAlign: "center",
+			marginLeft: isMobile? null : "65px",
 		}
 
 		const defaultImage = "/images/experienced.png"
 		const experienced_image = userdata !== undefined && userdata !== null && userdata.active_org !== undefined && userdata.active_org.image !== undefined && userdata.active_org.image !== null && userdata.active_org.image !== "" ? userdata.active_org.image : defaultImage
     return (
-				<div style={{width: 1000, margin: "auto", paddingBottom: 150, minHeight: 1500, marginTop: 50, }}>
+				<div style={{ margin: "auto", paddingBottom: 150, minHeight: 1500, marginTop: 50, }}>
 					{/*
 					<div style={{position: "fixed", bottom: 110, right: 110, display: "flex", }}>
 						<img src="/images/Arrow.png" style={{width: 250, height: "100%",}} />
@@ -342,7 +345,7 @@ const Welcome = (props) => {
 					*/}
 					{showWelcome === true ? 
 						<div>
-							<div style={{minWidth: 500, maxWidth: 500, margin: "auto", marginTop: isCloud ? "auto" : 20, }}>
+							{/* <div style={{minWidth: 500, maxWidth: 500, margin: "auto", marginTop: isCloud ? "auto" : 20, }}>
 								<Stepper 
 									activeStep={activeStep} 
 									color="primary"
@@ -375,8 +378,8 @@ const Welcome = (props) => {
 												)
 										})}
 								</Stepper>
-							</div>
-        			<Grid container spacing={2} style={{ padding: 0, maxWidth: 1000, minWidth: 1000, margin: "auto", }}>
+							</div> */}
+        			<Grid container spacing={1} style={{ padding: isMobile ? 10 : 0, maxWidth: isMobile ? null : 500, minWidth: isMobile ? null : 500, marginRight : "auto" , marginLeft: "auto" }}>
           		  <Grid item xs={window.location.href.includes("tab=2") ? 6 : 12}>
 									<div>
 											{/*
@@ -388,6 +391,8 @@ const Welcome = (props) => {
 											/>
 											*/}
 											<WelcomeForm2
+												isLoggedIn={isLoggedIn}
+												checkLogin={checkLogin}
 												userdata={userdata}
 												globalUrl={globalUrl}
 												discoveryWrapper={discoveryWrapper}
@@ -410,57 +415,22 @@ const Welcome = (props) => {
 											/>
 									</div>
           		  </Grid>
-								{frameworkData === undefined || window.location.href.includes("tab=1") || window.location.href.includes("tab=3") ? null :
-									<div style={{marginTop: 25, }}>
-										<Typography variant="h6" style={{textAlign: "center", marginBottom: 25, }}>
-											App Framework
-										</Typography>
-										<Fade>
-												<AppFramework
-													inputUsecase={inputUsecase}
-													frameworkData={frameworkData}
-													setFrameworkData={setFrameworkData}
-													selectedOption={"Draw"}
-													showOptions={false}
-													isLoaded={true}
-													isLoggedIn={true}
-													globalUrl={globalUrl}
-													size={0.78}
-													color={theme.palette.platformColor}
-													discoveryWrapper={discoveryWrapper}
-													setDiscoveryWrapper={setDiscoveryWrapper}
-													apps={apps}
-													inputUsecases={usecases}
-													setInputUsecases={setUsecases}
-												/>
-										</Fade>
-									</div>
-								}
 							</Grid>
 						</div>
 						: 
 						<Fade in={true}>
-							<div style={{maxWidth: 700, margin: "auto", marginTop: 50, }}>
-								{/*
-							<div style={{display:"flex"}}>
-								<ArrowBackIosIcon style={{color: "#9E9E9E",}} onClick={() => {
-										navigate("/login")
-									}}/>
-								<Typography variant="body1" style={{color: "#9E9E9E",textAlign: "center", marginBottom: 50, paddingRight: "366px"}} onClick={() => {
-										navigate("/login")
-									}}>
-								Back
-								</Typography>
-							</div>
-								*/}
-								<Typography variant="h4" style={{color: "#F1F1F1", textAlign: "center", marginTop: 50, }}>
+							<div style={{maxWidth: isMobile ? null : 590, margin: "auto", marginTop: 50,textAlign : isMobile ? "center" : null }}>
+								<Typography variant="h4" style={{color: "#F1F1F1", marginTop: 50, fontSize: 32}}>
 									Help us get to know you
 								</Typography>
-								<Typography variant="body1" style={{color: "#9E9E9E", textAlign: "center", marginBottom: 50,}}>
-									We will use this information to personalize your automation
+								<Typography variant="body1" style={{color: "#9E9E9E", marginTop: 12 ,marginBottom: 50, fontSize: 16}}>
+								Let us help you create a smoother journey.
 								</Typography>
-								<div style={{display: "flex", marginTop: 70, width: 700, margin: "auto",}}>
-									<div style={{border: "1px solid #806BFF", borderRadius: theme.palette.borderRadius, }}>
+								{/* <Typography variant="body1" style={{color: "#9E9E9E", marginTop: 12 ,marginBottom: 50, fontSize: 16}}>
+									We will use this information to personalize your automation
+								</Typography> */}
+								<div style={{display: isMobile ? null : "flex", marginTop: 70, width: isMobile ? 280 : 700, margin: "auto",}}>
+									<div style={{border: "2px solid #806BFF", borderRadius: theme.palette.borderRadius, }}>
 										<Card style={paperObject} onClick={() => {
 											if (isCloud) {
 													ReactGA.event({
@@ -472,6 +442,7 @@ const Welcome = (props) => {
 												//setActiveStep(1)
 											}
 
+											navigate("/welcome?tab=2")
 											setShowWelcome(true)
 										}}>
 											<CardActionArea style={actionObject}>
@@ -480,12 +451,13 @@ const Welcome = (props) => {
 													New to Gsoc2 
 												</Typography>
 												<Typography variant="body1" style={{marginTop: 10, color: "rgba(255,255,255,0.8)"}}>
-														Let us guide you for an easier experience
+													Let us guide you for an easier experience
 												</Typography>
 											</CardActionArea>
 										</Card>
 									</div>
-									<div style={{marginLeft: 25, marginRight: 25, }}>
+									<div style={{marginLeft: 25, marginRight: 25, 
+			marginTop: isMobile ? 30 : null, }}>
 										{/* <Typography style={{marginTop: 200, }}>
 											OR
 										</Typography> */}
@@ -502,7 +474,7 @@ const Welcome = (props) => {
 										navigate("/workflows?message=Skipped intro")
 									}}>
 										<CardActionArea style={actionObject}>
-											<img src={experienced_image} style={{padding: experienced_image === defaultImage ? 38 : 10, objectFit: "scale-down", minHeight: experienced_image === defaultImage ? 40 : 70, maxHeight: experienced_image === defaultImage ? 40 : 70, bordeRadius: theme.palette.borderRadius, }} />
+											<img src={experienced_image} style={{padding: experienced_image === defaultImage ? 2 : 10, objectFit: "scale-down", minHeight: experienced_image === defaultImage ? 40 : 70, maxHeight: experienced_image === defaultImage ? 40 : 70, bordeRadius: theme.palette.borderRadius*2, marginBottom: experienced_image === defaultImage ? 24 : 2 }} />
 											<Typography variant="h4" style={{color: "#F1F1F1"}}> 
 												Experienced 
 											</Typography>										
@@ -513,26 +485,20 @@ const Welcome = (props) => {
 									</Card>
 								</div>
 
-								{/*
-								<div style={{display: "flex", flexDirection: "row", }}>
+								
+								<div style={{ flexDirection: "row", textAlign: isMobile ? "center" : null, margin: isMobile ? "auto" : null }}>
 									<Button variant="contained" type="submit" fullWidth style={buttonStyle} onClick={() => {
-										navigate("/workflows?message=Skipped intro continue")
+										navigate(`/welcome?tab=2`)
+										setActiveStep(1)
+										setShowWelcome(true)
+										ReactGA.event({
+											category: "welcome",
+											action: "click_welcome_continue",
+											label: "",
+										})
 									}}>
 										Continue
 									</Button>
-								</div>
-								*/}
-
-								<div style={{margin: "auto", borderRadius: theme.palette.borderRadius, marginTop: 50, width: 200, overflow: "wrap", padding: 25, cursor: "pointer", }} onClick={() => {
-									if (window.drift !== undefined) {
-										window.drift.api.startInteraction({ interactionId: 340045 })
-									} else {
-										console.log("Couldn't find drift in window.drift and not .drift-open-chat with querySelector: ", window.drift)
-									}
-								}}>
-									<Typography variant="body1" style={{margin: "auto", textAlign: "center"}}>
-										Want a demo instead?
-									</Typography>
 								</div>
 							</div>
 						</Fade>
